requirements.txt
click>=8.1.7
rich>=13.7.0
pyyaml>=6.0.1